almost working sphere. float range check error
