+++ include/utils.h
@@ -3,11 +3,11 @@
 
 # include "minirt.h"
 
-void	init_a(t_amlight *list);
-void	init_c(t_camera *list);
-void	init_l(t_light *list);
-void	init_pl(t_plane *list);
-void	init_cy(t_cyl *list);
-void	init_sp(t_sphere *list);
-void	init_ent(t_entire *ent);
+// void	init_a(t_amlight *list);
+// void	init_c(t_camera *list);
+// void	init_l(t_light *list);
+// void	init_pl(t_plane *list);
+// void	init_cy(t_cyl *list);
+// void	init_sp(t_sphere *list);
+// void	init_ent(t_entire *ent);
 #endif
