+++ include/vector.h
@@ -13,20 +13,22 @@
 #ifndef VECTOR_H
 # define VECTOR_H
 
-#include "minirt.h";
+// #include "minirt.h"
 
-t_vector	*vector_init(float x, float y, float z);
-t_vector	*vecsubtraction(t_vector *vec1, t_vector *vec2);
-float		vec_length(t_vector *vec);
-void		vec_normalize(t_vector *vec);
-float		vec_product(t_vector *vec1, t_vector *vec2);
+// t_vector	*vector_init(float x, float y, float z);
+// t_vector	*vecsubtraction(t_vector *vec1, t_vector *vec2);
+// float		vec_length(t_vector *vec);
+// void		vec_normalize(t_vector *vec);
+// float		vec_product(t_vector *vec1, t_vector *vec2);
+// int			sphere_intersect(t_camera *camera, t_vector *ray, t_sphere *sphere);
+// void		ray_tracing(void *mlx, void *window, t_scene *scene);
 
-typedef struct s_vplane
-{
-	float	width;
-	float	height;
-	float	x_pixel;
-	float	y_pixel;
-}	t_vplane;
+// typedef struct s_vplane
+// {
+// 	float	width;
+// 	float	height;
+// 	float	x_pixel;
+// 	float	y_pixel;
+// }	t_vplane;
 
 #endif
