+++ src/parcing/free_and_to_struct1.c
@@ -6,8 +6,17 @@ void	int_range_checker(int *key, int value, int flag_range)
 		|| (flag_range == 2 && (value >= 0 && value <= 180)))
 		*key = value;
 	else
-		error(1);
+		error("int range check\n");
 }
+
+// void	float_range_checker(float *key, int value, int flag_range)
+// {
+// 	if ((flag_range == 1 && (value >= 0 && value <= 255))
+// 		|| (flag_range == 2 && (value >= 0 && value <= 180)))
+// 		*key = value;
+// 	else
+// 		error(1);
+// }
 
 void	free_a(char **split_str_2)
 {
@@ -62,7 +71,7 @@ t_camera	*to_struct_c(char **str)
 		return (NULL);
 	init_c(list);
 	count_of_split(str, 4);
-	int_range_checker(&(list->fov), ft_atoi(str[3]), 2);
+	float_range_checker(&(list->fov), ft_atoi(str[3]), 2);
 	split_str_1 = ft_split(str[1], ',');
 	split_str_2 = ft_split(str[2], ',');
 	count_of_split(split_str_1, 3);
